[test] test suite: move events parser to different files
